requirements.txt
chromadb>=0.3.27
sentence-transformers>=2.2.2
numpy>=1.24
openai>=1.0.0
tqdm>=4.64.0
python-dotenv>=0.21.0
pytest>=9.0.1